rename: adjust structure
