+++ src/components/TodoModal/TodoCreateModal/index.tsx
@@ -10,9 +10,9 @@ import useOutsideClick from '@hooks/useOutsideClick';
 import useVisibility from '@hooks/useVisibility';
 import { AxiosResponse } from 'axios';
 import { ChangeEvent, useEffect, useRef, useState } from 'react';
-import FileLinkSection from '../FileLinkSection';
-import GoalSection from '../GoalSection';
-import TitleSection from '../TitleSection';
+import FileLinkSection from '../components/FileLinkSection';
+import GoalSection from '../components/GoalSection';
+import TitleSection from '../components/TitleSection';
 
 interface TodoCreateModalProps {
   onClose: () => void;
