+++ src/components/TodoModal/TodoDetailModal/index.tsx
@@ -11,10 +11,10 @@ import useOutsideClick from '@hooks/useOutsideClick';
 import useVisibility from '@hooks/useVisibility';
 import { AxiosResponse } from 'axios';
 import { ChangeEvent, useRef, useState } from 'react';
-import FileLinkSection from '../FileLinkSection';
-import GoalSection from '../GoalSection';
-import StatusSection from '../StatusSection';
-import TitleSection from '../TitleSection';
+import FileLinkSection from '../components/FileLinkSection';
+import GoalSection from '../components/GoalSection';
+import TitleSection from '../components/TitleSection';
+import StatusSection from './StatusSection';
 import useTodoDetail from './useTodoDetail';
 
 export interface TodoDetailModalProps {
